+++ src/pages/SelectWeekPage/AdminDetailSection/ClosedDetail.tsx
