+++ src/pages/SelectWeekPage/AdminDetailSection/OpenDetail.tsx
