+++ src/pages/alba/ApplyPage/index.tsx
@@ -1,9 +1,9 @@
 import PageContainer from 'components/@commons/PageContainer';
-import SelectWeekPage from 'pages/admin/SelectWeekPage';
+import SelectWeekPage from 'pages/SelectWeekPage';
 import React from 'react';
 
 const ApplyPage = (): JSX.Element => {
-  return <SelectWeekPage />;
+  return <SelectWeekPage isAdmin={false} />;
 };
 
 export default ApplyPage;
