--- src/pages/alba/WeeklyStatusPage/index.tsx
@@ -1,8 +1,0 @@
-import SelectWeekPage from 'pages/SelectWeekPage';
-import React from 'react';
-
-const AlbaSelectWeekPage = (): JSX.Element => {
-  return <SelectWeekPage isAdmin={false} />;
-};
-
-export default AlbaSelectWeekPage;
